safety check: environments must be fresh
It is unsafe to instantiate the same environment multiple times with
different paths. The reason is that `Isabelle_System.init` will happily
succeed with an updated path, but discard it. This commit introduces a
check ensuring "freshness": If an environment is instantiated, it must
be fresh (i.e., has not been instantiated before) or the paths must
match. Users going through the high-level `Implementations` API will not
run into the second case, because it will always instantiate
environments in a fresh class loader.

Fixes #11.