remove scoped ML expressions
It turned out that forced resource safety through dependent types is not
the best idea. There are multiple problems:

* nesting scopes is really cumbersome (if it works at all)
* boilerplate because lambdas can't have a dependent type
* it doesn't work well on Scala 2.10

This commit retains resource safety "by convention": A handle to an
opaque ML term is passed in via a plain anonymous function, with the
assumption that the receiver doesn't pass it out by means of mutability
or other tricks.