prevent specs2 from finalizing the spec prematurely
This is a really odd one. On 3.8.6, this worked fine. On 3.8.7, this
started to fail on one machine (quad core). It worked fine on another
quad core machine.

My initial suspicion was that `afterAll` is executed after the body of
the spec has been evaluated, but not necessarily after all pending
examples have finished. This would lead to premature shutdown and
cancellation of pending futures.

It is still not clear whether that suspicion is accurate. I was unable
to reproduce the problem in a toy spec.

The `sequential` combinator makes all examples run sequentially. Still,
all futures are running in parallel. Exactly why this helps anyway is a
complete mystery to me.